% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexttab-package.R
\docType{package}
\name{contexttab-package}
\alias{contexttab}
\alias{contexttab-package}
\title{contexttab: ConTextTab, A Semantics-Aware Tabular In-Context Learner}
\description{
R bindings for the paper ConTextTab: A Semantics-Aware Tabular In-Context Learner. Provides functions for running the in-context learner (ICL) by SAP; sap-rpt-1-oss (formerly known under the name ConTextTab).
}
\author{
\strong{Maintainer}: frank hull \email{frankiethull@proton.me}

}
\keyword{internal}
