% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexttab-python.R
\name{install_contexttab_dependencies}
\alias{install_contexttab_dependencies}
\title{Install contexttab}
\usage{
install_contexttab_dependencies(envname = "contexttab", method = "auto", ...)
}
\arguments{
\item{envname}{virtual environment name}

\item{method}{method defaults to "auto"}

\item{...}{additional passes for \code{py_install}}
}
\value{
installs to "contexttab" env by default
}
\description{
Install contexttab
}
