% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexttab-pred.R
\name{predict.sap_rpt_oss.rpt.SAP_RPT_OSS_Regressor}
\alias{predict.sap_rpt_oss.rpt.SAP_RPT_OSS_Regressor}
\title{S3 prediction method for SAP_RPT_OSS_Regressor class}
\usage{
\method{predict}{sap_rpt_oss.rpt.SAP_RPT_OSS_Regressor}(fit, X, type = "raw", ...)
}
\arguments{
\item{fit}{a sap-rpt-oss fit}

\item{X}{predictors}

\item{type}{raw}

\item{...}{a placeholder}
}
\value{
predictions
}
\description{
S3 prediction method for SAP_RPT_OSS_Regressor class
}
