% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexttab-fit.R
\name{sap_rpt_oss_regressor_fit}
\alias{sap_rpt_oss_regressor_fit}
\title{Fit SAP RPT OOS (fka ConTextTab) Regressor}
\usage{
sap_rpt_oss_regressor_fit(X, y, ...)
}
\arguments{
\item{X}{a set of predictors}

\item{y}{a label}

\item{...}{additional parameters to pass to the fitting engine}
}
\value{
a fitted model
}
\description{
Fit SAP RPT OOS (fka ConTextTab) Regressor
}
